backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
google-generativeai==0.5.4
python-dotenv==1.0.1
whoosh==2.7.4
python-multipart==0.0.9
pydantic==2.7.1
